frontend: refatorado código para melhor leitura e criado componente de crud para padronização de layout na visão de tabela

+++ client/src/app/pages/admin/administrative/profiles/profiles.component.ts
@@ -1,13 +1,19 @@
 import { CommonModule } from '@angular/common';
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCardModule } from '@angular/material/card';
-import { Router } from '@angular/router';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatSort } from '@angular/material/sort';
+import { MatTableDataSource } from '@angular/material/table';
+import { ConfirmService } from 'app/components/confirm/confirm.service';
 import { LoadingService } from 'app/components/loading/loading.service';
+import { ModalService } from 'app/components/modal/modal.service';
 import { NotFoundRegisterComponent } from 'app/components/not-found-register/not-found-register.component';
 import { ToastService } from 'app/components/toast/toast.service';
 import { Profile } from 'app/model/Profile';
 import { MESSAGES } from 'app/utils/messages';
+import { CrudComponent } from '../../../../components/crud/crud.component';
+import { ProfileComponent } from './profile/profile.component';
 import { ProfilesService } from './profiles.service';
 
 @Component({
@@ -20,23 +26,29 @@ import { ProfilesService } from './profiles.service';
     MatButtonModule,
     NotFoundRegisterComponent,
     CommonModule,
+    CrudComponent,
   ],
 })
 export class ProfilesComponent implements OnInit {
   profiles: Profile[] = [];
-  displayedColumns: string[] = ['name', 'description', 'status', 'actions'];
-  columnDefinitions = {
-    name: 'Nome',
-    description: 'Descrição',
-    status: 'Situação',
-    actions: 'Ações',
-  };
+  rendering: boolean = true;
+  dataSourceMat = new MatTableDataSource<Profile>(this.profiles);
+  @ViewChild(MatPaginator) paginator!: MatPaginator;
+  @ViewChild(MatSort) sort!: MatSort;
+
+  columnDefinitions = [
+    { key: 'status', header: 'Situação', type: 'boolean' },
+    { key: 'name', header: 'Cargo', type: 'string' },
+    { key: 'description', header: 'Descrição', type: 'string' },
+    { key: 'updated_at', header: 'Última atalização', type: 'datetime' },
+  ];
 
   constructor(
-    private router: Router,
-    private profilesService: ProfilesService,
+    private modal: ModalService,
+    private confirmModal: ConfirmService,
     private toast: ToastService,
     private loading: LoadingService,
+    private profilesService: ProfilesService,
   ) {}
 
   ngOnInit() {
@@ -44,10 +56,13 @@ export class ProfilesComponent implements OnInit {
   }
 
   loadProfiles = () => {
-    this.loading.show();
     this.profilesService.getProfiles().subscribe({
       next: (profiles) => {
         this.profiles = profiles;
+        this.dataSourceMat.data = this.profiles;
+        this.dataSourceMat.paginator = this.paginator;
+        this.dataSourceMat.sort = this.sort;
+        this.rendering = false;
       },
       error: () => {
         this.loading.hide();
@@ -57,22 +72,81 @@ export class ProfilesComponent implements OnInit {
     });
   };
 
-  addNewProfile = (): void => {
-    console.log('addNewProfile');
-  };
+  addNewProfile = () => {
+    const modal = this.modal.openModal(
+      `modal-${Math.random()}`,
+      ProfileComponent,
+      'Adicionando um novo perfil',
+      true,
+      true,
+    );
 
-  editProfile = (profile: Profile): void => {
-    console.log('editProfile', profile);
-  };
-
-  deleteProfile = (profile: Profile): void => {
-    this.profilesService.deleteProfile(profile.id).subscribe({
-      next: () => {
-        this.toast.openSuccess('Perfil excluído com sucesso!');
+    modal.afterClosed().subscribe((result) => {
+      if (result) {
         this.loadProfiles();
+      }
+    });
+  };
+
+  editProfile = (profile: Profile) => {
+    const modal = this.modal.openModal(
+      `modal-${Math.random()}`,
+      ProfileComponent,
+      'Editando um perfil',
+      true,
+      true,
+      { profile },
+    );
+
+    modal.afterClosed().subscribe((result) => {
+      if (result) {
+        this.loadProfiles();
+      }
+    });
+  };
+
+  deleteProfile = (profile: Profile) => {
+    const modal = this.confirmModal.openConfirm(
+      'Confirmar exclusão',
+      `Tem certeza que deseja excluir o perfil ${profile.name}?`,
+      'Excluir',
+      'Cancelar',
+    );
+
+    modal.afterClosed().subscribe((result: Profile) => {
+      if (result) {
+        this.profilesService.deleteProfile(profile.id).subscribe({
+          next: () => this.toast.openSuccess(MESSAGES.DELETE_SUCCESS),
+          error: () => {
+            this.loading.hide();
+            this.toast.openError(MESSAGES.DELETE_ERROR);
+          },
+          complete: () => {
+            this.loadProfiles();
+            this.loading.hide();
+          },
+        });
+      }
+    });
+  };
+
+  toggleStatus = (profile: Profile) => {
+    const updatedStatus = !profile.status;
+    profile.status = updatedStatus;
+
+    this.profilesService.updatedStatus(profile.id, updatedStatus).subscribe({
+      next: () => {
+        this.toast.openSuccess(
+          `Perfil ${updatedStatus ? 'ativado' : 'desativado'} com sucesso!`,
+        );
       },
       error: () => {
-        this.toast.openError('Erro ao excluir o perfil!');
+        this.loading.hide();
+        this.toast.openError(MESSAGES.UPDATE_ERROR);
+      },
+      complete: () => {
+        this.loadProfiles();
+        this.loading.hide();
       },
     });
   };
